fixing problems
