+++ 0001/list-questions/src/main/java/ru/otus/Main.java
@@ -1,10 +1,8 @@
 package ru.otus;
 
 import org.springframework.context.support.ClassPathXmlApplicationContext;
-import ru.otus.dao.QuestionDao;
-import ru.otus.dto.QuestionItem;
+import ru.otus.service.QuestionService;
 
-import java.util.List;
 
 
 
@@ -13,13 +11,9 @@ public class Main {
     public static void main(String[] args) {
 
         ClassPathXmlApplicationContext context = new ClassPathXmlApplicationContext("/spring-context.xml");
-        QuestionDao questionDao = context.getBean(QuestionDao.class);
-        List<QuestionItem> questionItems = questionDao.getQuestionItems();
+        QuestionService questionService = context.getBean(QuestionService.class);
 
-        System.out.println("--------questions-----------");
-        for(QuestionItem questionItem : questionItems){
-            System.out.println(questionItem);
-            System.out.println("-------------------");
-        }
+        questionService.printQuestions();
+
     }
 }
