--- 0001/list-questions/src/main/java/ru/otus/dao/LoadDataDao.java
@@ -1,7 +1,0 @@
-package ru.otus.dao;
-
-import java.util.List;
-
-public interface LoadDataDao {
-    public List<String[]> getData();
-}
