+++ 0001/list-questions/src/test/java/dao/QuestionImplTest.java
@@ -2,29 +2,17 @@ package dao;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-import ru.otus.dao.LoadDataDao;
-import ru.otus.dao.QuestionImpl;
+import ru.otus.dao.QuestionDaoImpl;
 import ru.otus.dto.QuestionItem;
-
-import java.util.ArrayList;
 import java.util.List;
 
 public class QuestionImplTest {
-    public class LoadDataImpl implements LoadDataDao {
 
-        @Override
-        public List<String[]> getData(){
-            return new ArrayList<>(){{
-                add(new String[]{"2+1","3","5"});
-                add(new String[]{"2+2","4","5"});
-            }};
-        }
-    }
     @Test
     public void testWork(){
 
-        LoadDataImpl loadDataImpl = new LoadDataImpl();
-        QuestionImpl questionImpl = new QuestionImpl(loadDataImpl);
+
+        QuestionDaoImpl questionImpl = new QuestionDaoImpl("data.csv");
 
         Assertions.assertNotNull(questionImpl);
