+++ 0001/list-questions/src/test/resources/data.csv
@@ -1,0 +1,2 @@
+6+3,9,8,7,6,5,4,3
+7+1,8,6,5,4,3,1
